# Conflicts:
#	ChronoLoop/ChronoLoop/Actions/TeleportAction.hpp
#	ChronoLoop/ChronoLoop/Objects/Component.h
#	ChronoLoop/ChronoLoop/Physics/Physics.cpp

+++ ChronoLoop/ChronoLoop/Core/Timeline.h
@@ -55,13 +55,13 @@ namespace Epoch {
 	};
 
 	struct ObjectLifeTime {
-		int mBirth = -1;
-		int mDeath = INT32_MAX;
+		unsigned int mBirth = -1;
+		unsigned int mDeath = INT32_MAX;
 	};
 
 	class Timeline {
 		//Where we are at in the timeline
-		int mCurrentGameTimeIndx = 0;
+		unsigned int mCurrentGameTimeIndx = 0;
 		std::vector<unsigned int> mSnaptimes;
 		std::unordered_map<unsigned int, Snapshot*> mSnapshots;		//The key will be the time they were taken (mSnapTimes)
 		std::unordered_map<unsigned short, BaseObject*> mLiveObjects;
@@ -82,8 +82,8 @@ namespace Epoch {
 		SnapInfo* GenerateSnapInfo(BaseObject* _object, SnapInfo* _info);							//Error check agianst the BaseObject* if it is null or not
 		Snapshot* GenerateSnapShot(unsigned int _time, std::vector<BaseObject*> & _clones);
 		SnapInfoPlayer * GenerateSnapInfoPlayer();
-		int GetCurrentGameTimeIndx() { return mCurrentGameTimeIndx; }
-		unsigned int GetTotalSnaps() { return mSnapshots.size(); };
+		unsigned int GetCurrentGameTimeIndx() { return mCurrentGameTimeIndx; }
+		unsigned int GetTotalSnaps() { return (unsigned int)mSnapshots.size(); };
 		void HotFixResetLevel();
 		bool RewindMakeClone(unsigned int _snaptime);
 		bool RewindNoClone(unsigned int _snaptime, unsigned short _id1, unsigned short _id2, unsigned short _id3);
@@ -95,7 +95,7 @@ namespace Epoch {
 		void SetCloneDeathTime(unsigned short _id1, unsigned short _id2, unsigned short _id3);
 		void SetComponent(SnapComponent* _destComp, BaseObject* _obj, SnapInfo* _destInfo);
 		void SetCurrentGameTimeIndx(int _time) { mCurrentGameTimeIndx = _time; };
-		void UpdateCloneInterpolators(unsigned short _cloneid, SnapInfo* _currSnap, float _currTime);
+		void UpdateCloneInterpolators(unsigned short _cloneid, SnapInfo* _currSnap, unsigned int _currTime);
 	};
 
 }
